Instrument and fix launch.py, such that it respects faction settings.
Before this fix, non-blank ships would almost never be launched.

Instead, blank ships would be selected, upgraded and launched, which
is problematic insofar as the ship upgrade routine does not appear to
respect class restrictions.

After this fix, it is again possible to configure whether a faction
should always launch non-blank ships or if there should be a certain
chance to launch non-standard, upgraded .blank ships.

It is worth noting that the ship upgrade routine was not changed
in this commit.