Experiment with custom upgrade rules.
For now, non-blank pirate ships in Troy will be downgraded if
the player is flying a Tarsus, because we're assuming that as
the player progresses, he'll sell his tarsus.begin for something
slightly less sucky.

Note that the TarsusMk2 is currently considered a tarsus as
far as the custom upgrade rule is considered.